+++ src/main/java/com/total/received/appointments/GetTotalReceivedAppointments/model/CustomResponse.java
@@ -1,17 +1,19 @@
 package com.total.received.appointments.GetTotalReceivedAppointments.model;
 
+import java.math.BigInteger;
+
 public class CustomResponse {
 
-    private int totalRecResponse;
+    private BigInteger totalRecResponse;
 
     public CustomResponse() {
     }
 
-    public int getTotalRecResponse() {
+    public BigInteger getTotalRecResponse() {
         return totalRecResponse;
     }
 
-    public void setTotalRecResponse(int totalRecResponse) {
+    public void setTotalRecResponse(BigInteger totalRecResponse) {
         this.totalRecResponse = totalRecResponse;
     }
 }
